Range parse refactor. Decouple it from the textbuf struct by using a union type that represent generic Range. (#24)
So, instead of passing a context to the parse function with the textbuf to get from there the data needed to fill the actual range values, use a RangeAddr type that encode those as current, end, search, previous, etc. and the actual values are set afterwards with a specific textbuf method.

Reviewed-on: https://codeberg.org/nsm/ahre/pulls/24
Co-authored-by: Nicolas San Martin <nsanmartin@dc.uba.ar>
Co-committed-by: Nicolas San Martin <nsanmartin@dc.uba.ar>