Replace the [static 1] pointer parameter for [_1_] with _1_ == 1
[static 1] does not make the compiler to check that the function
does not receive a nullptr. Instead, it assumes this and causes
undefined behaviour if it does. Instead, just [1] is, according
the standard, the same as a plain pointer (de to array decay),
but the compilers (ggc and clang) seem to be more strict and add
some checks. It appear (se the linked discussion) also that there
is some initiative to make [1] the compiler check and diverge from
array decay.

https://lwn.net/SubscriberLink/1046840/3eb9029084cc9e1e/